Renamed upperGateIn/lowerGateIn to upperLayerIn/lowerLayerIn and upperGateOut/lowerGateOut to upperLayerOut/lowerLayerOut.
Renamed INic to IWirelessNic. Allow the upperControlOut and the lowerControlOut gates to be unconnected.

The IWirelessNic moduleinterface in INET does not have control gates. Drop the control messages that were sent through these gates if unconnected.

All this is necessary to be able to interoperate with INET. (Levente)

Fixed erasing of elements in Decider80211MultiChannel.cc.